fix: move EditableFieldAdapter to adapter sub-package, and extract all ViewHolder in dedicated classes
