+++ occtax/src/main/java/fr/geonature/occtax/ui/input/counting/EditCountingMetadataFragment.kt
@@ -31,7 +31,7 @@ import fr.geonature.compat.os.getParcelableArrayCompat
 import fr.geonature.compat.os.getParcelableCompat
 import fr.geonature.occtax.R
 import fr.geonature.occtax.features.nomenclature.domain.EditableField
-import fr.geonature.occtax.features.nomenclature.presentation.EditableFieldAdapter
+import fr.geonature.occtax.features.nomenclature.presentation.adapter.EditableFieldAdapter
 import fr.geonature.occtax.features.nomenclature.presentation.NomenclatureViewModel
 import fr.geonature.occtax.features.nomenclature.presentation.PropertyValueModel
 import fr.geonature.occtax.features.record.domain.CountingRecord
